Add solutions to Day 3 (both parts)
Part one was fairly trivial.

Part two is way more interesting.
Since now we're looking at the whole dataset in memory (instead of only each line in isolation as before), we can see that deciding if we have more 0's or 1's is equivalent of doing a binary search in the sorted vector and check how many elements we have on each side of it. Then depending on each rating we're computing, we go into the side with more or less elements.

Another possible approach would be to create something like:
struct BitTree {
  zeroSubTree: BitTree,
  zeroCount: u32,
  oneSubTree: BitTree,
  oneCount: u32,
}

And as you read the input, add the numbers encoded in this bit tree. Then it's trivial after that to traverse it to find any of the ratings. This one might be easier to understand than wrapping the head around the binary searches, but it's essentially the same thing... just encoded in a different way =)