[5/N] Change Solver trait methods to receive a BufReader provider
This fix a bug when calling solver.solve_both(), because some solvers may
require two passes over the input or may have separate implementations
for each part and requires reading the input twice.

Now we pass a closure that return a new BufReader when necessary.

(We could also have read the whole content of the input in memory and keep reusing that, would be simpler. But I'll keep doing the input streaming version
since I was doing this since the beginning of the AdventOfCode 2021.)